requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
mysql-connector-python>=8.0.0
flask>=3.0.0
flask-cors>=4.0.0
playwright>=1.40.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
bcrypt>=4.0.1
gunicorn>=21.2.0
